Added validaton before delete MainDocumentTemplate

+++ src/main/java/ch/swissbytes/Service/business/mainDocument/MainDocumentService.java
@@ -4,8 +4,8 @@ package ch.swissbytes.Service.business.mainDocument;
 import ch.swissbytes.Service.business.Service;
 import ch.swissbytes.domain.model.entities.AttachmentMainDocumentEntity;
 import ch.swissbytes.domain.model.entities.MainDocumentEntity;
-import ch.swissbytes.domain.model.entities.TextSnippetEntity;
 import ch.swissbytes.domain.types.StatusEnum;
+import org.omnifaces.util.Messages;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
@@ -35,14 +35,18 @@ public class MainDocumentService extends Service<MainDocumentEntity> implements 
 
     @Transactional
     public void delete(MainDocumentEntity entity) {
-        entity.setStatus(StatusEnum.DELETED);
-        entity.setLastUpdate(new Date());
-        if (entity.getAttachmentMainDocument() != null) {
-            AttachmentMainDocumentEntity toDelete = entity.getAttachmentMainDocument();
-            entity.setAttachmentMainDocument(null);
-            attachmentService.delete(toDelete);
+        if (canDeleteMainDocumentSelected(entity)) {
+            entity.setStatus(StatusEnum.DELETED);
+            entity.setLastUpdate(new Date());
+            if (entity.getAttachmentMainDocument() != null) {
+                AttachmentMainDocumentEntity toDelete = entity.getAttachmentMainDocument();
+                entity.setAttachmentMainDocument(null);
+                attachmentService.delete(toDelete);
+            }
+            dao.update(entity);
+        } else {
+            Messages.addFlashGlobalError("Can not delete " + entity.getCode() + " because it is already being used");
         }
-        dao.update(entity);
     }
 
     @Transactional
@@ -115,5 +119,10 @@ public class MainDocumentService extends Service<MainDocumentEntity> implements 
     public MainDocumentEntity findByProjectIdAndCode(Long projectId, String code) {
         List<MainDocumentEntity> list = dao.findByProjectIdAndCode(projectId, code);
         return list.isEmpty() ? null : list.get(0);
+    }
+
+    public boolean canDeleteMainDocumentSelected(MainDocumentEntity entity) {
+        Long attachId = entity.getAttachmentMainDocument() != null ? entity.getAttachmentMainDocument().getId() : -1l;
+        return dao.findProjectDocumentByAttachmentIdOrMainDocumentId(attachId, entity.getId()).isEmpty() ? true : false;
     }
 }
